refactore: reorganize api package and export public entities
